move getbakerlist to index page
